Refactored BulkEditor into multiple classes
Split code of BulkEditor plugin into multiple classes to make it more
readable and easier to extend/change
