+++ plugins/org.eclipse.fordiac.ide.bulkeditor/plugin.properties
@@ -59,5 +59,7 @@ Save=Save
 Discard=Discard
 Cancel=Cancel
 
+UnexpectedValue=Unexpected value: {0}
+
 context.bulkEditing.description=4diac IDE Bulk Editing Context
 context.bulkEditing.name=Bulk Editing
